feat: arrange the forms
